chore: moved MenuRegistry to internal package
Removed Serializable from MenuConfiguration.
