+++ flow-server/src/main/java/com/vaadin/flow/router/internal/DefaultRouteResolver.java
@@ -16,7 +16,6 @@
 package com.vaadin.flow.router.internal;
 
 import java.util.Collections;
-import java.util.List;
 
 import org.slf4j.LoggerFactory;
 
@@ -25,11 +24,8 @@ import com.vaadin.flow.router.NavigationState;
 import com.vaadin.flow.router.NavigationStateBuilder;
 import com.vaadin.flow.router.NotFoundException;
 import com.vaadin.flow.router.RouteResolver;
-import com.vaadin.flow.router.RouterLayout;
 import com.vaadin.flow.server.RouteRegistry;
-import com.vaadin.flow.server.VaadinSession;
-import com.vaadin.flow.server.menu.AvailableViewInfo;
-import com.vaadin.flow.server.menu.MenuRegistry;
+import com.vaadin.flow.internal.menu.MenuRegistry;
 
 /**
  * Default implementation of the {@link RouteResolver} interface.
