+++ flow-server/src/test/java/com/vaadin/flow/router/DefaultRouteResolverTest.java
@@ -34,7 +34,7 @@ import com.vaadin.flow.router.internal.DefaultRouteResolver;
 import com.vaadin.flow.router.internal.ResolveRequest;
 import com.vaadin.flow.server.InvalidRouteConfigurationException;
 import com.vaadin.flow.server.RouteRegistry;
-import com.vaadin.flow.server.menu.MenuRegistry;
+import com.vaadin.flow.internal.menu.MenuRegistry;
 
 public class DefaultRouteResolverTest extends RoutingTestBase {
