+++ flow-server/src/test/java/com/vaadin/flow/server/menu/MenuConfigurationTest.java
@@ -51,7 +51,7 @@ import com.vaadin.flow.server.VaadinSession;
 import com.vaadin.flow.server.startup.ApplicationRouteRegistry;
 
 import static com.vaadin.flow.server.frontend.FrontendUtils.GENERATED;
-import static com.vaadin.flow.server.menu.MenuRegistry.FILE_ROUTES_JSON_NAME;
+import static com.vaadin.flow.internal.menu.MenuRegistry.FILE_ROUTES_JSON_NAME;
 
 @NotThreadSafe
 public class MenuConfigurationTest {
